Added Background/Event Page with a ticker to check for issues.
Officially it remains a "Background Page" for now because I couldn't get
the Chrome Alarm system to work. Must be a trivial overlook, which
should be fixed in the future.

I needed to add additional Permissions to the extension so that the REST
API URL could be reached without violating "same origin policy". Uses
the Badge feature to show how many issues returned by the JQL (JIRA
Query Language). Right now the query is hard-coded, but in the future
should be configurable in the options.

More feature addition comments.