Ability to set Announcement Banners & started a README.md
Experimented with permissions for accessing the Cookie, but found a way
to send the POST request without any authentication token.

Updated the regex for searching for issue ID (added an EOL).

Changed the color of the Badge over the icon. Also it now shows an
exclamation when unable to retrieve the list and doesn't show anything
when the length is 0.