Restructured code
Separated out the analytics code, but still have it directly embedded in
background.js while I ponder using a background.html page. It would be
beneficial for the analytics code plus for the use of libraries such as
jQuery... but I'm not convinced yet for using it in my background page -
I think I will test to see if when it is an Event Page, if all the
resources are loaded when the Event Page is used.

Moved the functions from the Popup into background.js and making use of
extension messages to communicate between them.

Added jslint predefined global variables.

Should I place all html files into a ./views/ directory? We'll see!