Fixed click trigger for search & function call on interval
Only the first recurring query for the badge was being called.

The click trigger for the search button was not working because no
Project Prefix existed.

Customized the project data that is saved on extension launch.

Also changed the search "Go" to a button (was a submit input element
before).