Added customizable content script which is injected into pages
Added a text area where Javascript can be saved which will then be
injected and executed on each page load. Considering adding the option
to select at which point during the tab load the script will be
executed.

The injected script (jira-customizations) is not complete yet as I
realized that access to the jQuery object loaded by Atalassian is not
available inside content scripts. In fact, only the DOM is shared
amongst the page vs. content scripts.

Attempted to add a separate option field for loading an external
Javascript file, but I wasn't able to get the extension to allow it. At
this point, the conclusion is that it can't be done this way
programmatically (only through the use of the "content_scripts" in the
manifest file).

Also started adding additional information inside the popup, though it
is still incomplete.