Injects jQuery before running the user-defined Javascript
Will most likely make the "default value for the Description field" its
own customizable field in the options page since this is a commonly
desired functionality.

Don't forget to add a note in the Options page to indicate that there
will be a certain version of jQuery available for the script and the
option to choose when the script is executed.