Added a mediocre, generic-purpose AJAX function
I haven't updated all the AJAX calls to make use of this function just
yet even though it's been tested and is stable.

Noticed an empty object is now being sent to the crxMessage() function
from the popup so no message is ever processed. Don't know when this
started happening but will be looked at next.